+++ hybrid/src/timeseries_database/timeseries_sql_rewrite.rs
@@ -266,7 +266,7 @@ impl TimeSeriesQueryToSQLTransformer<'_> {
             } else {
                 panic!("Should never happen");
             };
-            let grp_value = if let AnyValue::Int32(grp_value) = grp {
+            let grp_value = if let AnyValue::Int64(grp_value) = grp {
                 grp_value
             } else {
                 panic!("Should never happen");
@@ -774,7 +774,7 @@ mod tests {
                                                 },
                                                 DataFrame::new(vec![
                                                     Series::new("ts_external_id_1", ["id1"]),
-                                                    Series::new("grouping_col_0", [0i32]),
+                                                    Series::new("grouping_col_0", [0i64]),
                                                 ])
                                                 .unwrap(),
                                                 "grouping_col_0".to_string(),
@@ -814,7 +814,7 @@ mod tests {
                                                 },
                                                 DataFrame::new(vec![
                                                     Series::new("ts_external_id_2", ["id2"]),
-                                                    Series::new("grouping_col_0", [1i32]),
+                                                    Series::new("grouping_col_0", [1i64]),
                                                 ])
                                                 .unwrap(),
                                                 "grouping_col_0".to_string(),
